requirements.txt
# MCP Base Server Dependencies
# This file lists dependencies for the mcp-base server itself
# (the server construction assistant)

# FastMCP - MCP server framework
fastmcp>=2.0.0

# ASGI server for HTTP transport
uvicorn[standard]>=0.30.0

# ASGI framework
starlette>=0.38.0

# Template rendering
jinja2>=3.1.0

# HTTP client
httpx>=0.27.0

# Async utilities
anyio>=4.0.0
